+++ lib/squadron_local_worker.dart
@@ -1,5 +1,5 @@
 export 'squadron_service.dart';
 
-export 'src/channel.dart';
+export 'src/channel.dart' hide PipeExtension;
 export 'src/local_worker.dart';
 export 'src/local_worker_client.dart';