+++ lib/src/xplat/_stream_wrapper.dart
@@ -12,16 +12,15 @@ class StreamWrapper<T> {
   /// operations will be initiated by sending the [streamRequest] to the worker using [postMethod]. This will not be
   /// done before a listener is attached to this instance's [stream] property.
   StreamWrapper(WorkerRequest streamRequest,
-      {required PostMethod postMethod,
+      {required PostRequest postMethod,
       required Stream messages,
       required SquadronCallback onDone,
-      CancellationToken? token,
-      required bool inspectRequest})
+      CancellationToken? token})
       : _streamRequest = streamRequest,
-        _postMethod = postMethod,
+        _postRequest = postMethod,
         _messages = messages,
-        _token = token,
-        _inspectRequest = inspectRequest {
+        _token = token {
+    _handle = _process;
     _controller = StreamController<T>(
       onListen: _onListen,
       onPause: _onPause,
@@ -40,15 +39,25 @@ class StreamWrapper<T> {
 
   final Stream<dynamic> _messages;
   final WorkerRequest _streamRequest;
-  final PostMethod _postMethod;
+  final PostRequest _postRequest;
   final CancellationToken? _token;
-  final bool _inspectRequest;
 
   late final StreamController<T> _controller;
+
   int _paused = 0;
+  late void Function(WorkerResponse) _handle;
 
-  void _canceller() => _postMethod(WorkerRequestImpl.cancel(_token!), false);
+  void _canceller() => _postRequest(WorkerRequestImpl.cancel(_token!));
 
+  List<WorkerResponse>? _buffer;
+
+  /// Bufferize worker responses (paused stream).
+  void _bufferize(WorkerResponse res) {
+    final buffer = (_buffer ??= <WorkerResponse>[]);
+    buffer.add(res);
+  }
+
+  /// Forward worker responses (non-paused stream).
   void _process(WorkerResponse res) {
     final error = res.error;
     if (error != null) {
@@ -58,61 +67,49 @@ class StreamWrapper<T> {
     }
   }
 
-  /// Buffered storing worker responses while the stream is paused.
-  final _buffer = <WorkerResponse>[];
-
-  /// Will eventually store the stream ID received from the worker. If the stream is cancelled on the client side,
-  /// the stream ID will be sent to the worker so it has a chance to stop streaming from the worker service.
+  /// Will eventually store the stream ID received from the worker. If the stream is cancelled on the
+  /// client side, the stream ID will be sent to the worker so it has a chance to stop streaming from
+  /// the worker service.
   final _streamId = Completer<int>();
 
   void _onListen() {
     _messages.listen(
       (message) {
         final res = message as List;
-        if (!res.unwrapResponse()) {
+        if (!res.unwrapResponse() || _controller.isClosed) {
           return;
         }
 
-        if (_controller.isClosed) return;
         final cancelException = _token?.exception;
         if (cancelException != null) {
           if (!_streamId.isCompleted) {
             _streamId.complete(-1);
           }
           _controller.addError(cancelException, cancelException.stackTrace);
-          _buffer.clear();
+          _buffer?.clear();
           _controller.close();
-          return;
-        }
-
-        if (res.endOfStream) {
+        } else if (res.endOfStream) {
           _controller.close();
-          return;
-        }
-        if (!_streamId.isCompleted) {
-          // The first message received from the worker contains the stream ID. If the stream is cancelled on
-          // the client side, the stream from the worker context should also be cancelled by sending a
-          // WorkerRequest.cancelStream with this stream id.
+        } else if (!_streamId.isCompleted) {
+          // The first message received from the worker contains the stream ID. If the stream
+          // is cancelled on the client side, the stream from the worker context should also
+          // be cancelled by sending a WorkerRequest.cancelStream with this stream id.
           _streamId.complete(res.result);
-          return;
-        }
-        if (_paused > 0) {
-          _buffer.add(res);
         } else {
-          _process(res);
+          _handle(res);
         }
       },
       cancelOnError: false,
     );
     // initiate streaming operations now!
-    _postMethod(_streamRequest, _inspectRequest);
+    _postRequest(_streamRequest);
   }
 
   Future _onCancel() async {
     final streamId = await _streamId.future;
     // notify the worker that the streaming operation has been cancelled
-    _postMethod(WorkerRequestImpl.cancelStream(streamId), false);
-    _buffer.clear();
+    _postRequest(WorkerRequestImpl.cancelStream(streamId));
+    _buffer?.clear();
     _controller.close();
   }
 
@@ -120,6 +117,9 @@ class StreamWrapper<T> {
     // per Dart documentation: https://api.dart.dev/dart-async/StreamSubscription/pause.html
     // If the subscription is paused more than once, an equal number of resumes must be performed to resume the
     // stream.
+    if (_paused == 0) {
+      _handle = _bufferize;
+    }
     _paused++;
   }
 
@@ -134,10 +134,12 @@ class StreamWrapper<T> {
       // per Dart documentation: https://api.dart.dev/dart-async/StreamSubscription/resume.html
       // When all previously calls to pause have been matched by a calls to resume, [...] the stream subscription
       // may emit events again.
-      for (var e in _buffer) {
-        _process(e);
+      _handle = _process;
+      final buffer = _buffer;
+      if (buffer != null) {
+        buffer.forEach(_process);
+        buffer.clear();
       }
-      _buffer.clear();
     }
   }
 }